pyproject.toml
[tool.poetry]
name = "taxogen"
version = "0.1.0"
description = ""
authors = ["Cezar Sas <cezar.sas@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.21.4"
scipy = "^1.7.3"
scikit-learn = "0.24.2"
PyYAML = "^6.0"


[tool.poetry.dev-dependencies]
2to3 = "^1.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
